Update ContactList.java
This next line will show the error message after customer presses enter once. If not the user pushes enter twice in the last name field and then the error pops up. Try the original code and see for yourself when you do not enter a last name it requires two presses of the enter button to show error. If you have any other questions ask me in the morning. And also the output should show an example of the  "Exception course: If the user leaves the last name blank, then the person will not be entered into the contact list." as it doesn't currently show it.